move recursive translation onto the strategy
